// bk/src/types/flow.ts
export interface FormFieldOption {
  value: string | number;
  label: string;
}

export interface QueryParameter {
  key: string;
  label: string;
  type: 'string' | 'number' | 'date';
  required: boolean;
}

export interface CustomQuery {
  name: string;
  description: string;
  targetEndpoint: string;
  isCatalog: boolean;
  parameters: QueryParameter[];
  locked?: boolean;
  mockResponse?: any; // For development/testing
}